Jump on top of platform from below it
The player can now jump on top of platform from below it.

If player object's rigidbody velocity on Y-axis is greater than 1f, then method is called which casts capsule cast and if it returns collider its gameObject's tag gets checked and if it is called "Platform": its isTrigger variable toggles itself to true.